+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+### Features
+* **@dlr-eoc/services-map-state**
+ - Allow to use `projjson` in `IProjDef`.
+ - Export `EPSG_3035_Def`
+ - Export function `getProj4Defs(projDef: IProjDef)`, `adjustBBoxAxis(bbox: [number, number, number, number], axis: AxisType)` and `adjustBBoxAxisToEnu(bbox: [number, number, number, number], proj?: TepsgCode | IProjDef, axis?: AxisType)` for convert to `enu`. Use this for `nativeBbox` get from wms 1.3.0. This tries to fixes native bbox axis from wms 1.3.0.
+
+* **@dlr-eoc/map-ol:**
+ - Use `projjson` or `proj4js` to create projection if defined on `IProjDef`.
+
+ 
+### Bug Fixes
+* **@dlr-eoc/ngx-ukis-ui-clarity:**
+ - Can zoom to `nativeBbox` for `LayerentryComponent` and `LayerentryGroupComponent`.
+
+* **@dlr-eoc/map-ol:**
+ - Transform extent in `setLayerExtentAfterProjection` if nativeBbox.epsg is not new epsg or try to use old extent.
+
+
 # [16.0.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v16.0.0) (2026-05-29) (update dependencies, new library ngx-ukis-ui-clarity and ngx-ukis-utilities, replace core-ui)
 
 ### Breaking Changes